RE-KERNEL-HUMAN-01-R2: repair P1 Research classifier defects
